refactor: move ComposableNode to descriptions/, split container/lcn files
- Add roscope/entities/descriptions/composable_node.py matching official
  launch_ros.descriptions.ComposableNode: plain description class (not
  Action), properties node_plugin/node_name/node_namespace, extra_arguments
  support, parse() classmethod with deferred Node import to break circular dep
- Split roscope/entities/actions/node.py into three files matching official:
  node.py (Node, LifecycleNode), composable_node_container.py
  (ComposableNodeContainer + _resolve_plugin helpers), load_composable_nodes.py
  (LoadComposableNodes)
- Condition filtering moved to ComposableNodeContainer.execute() (matching
  official), replacing eager parse-time filtering
- Update all imports in resolver.py, graph.py, tests/
