+++ roscope/entities/actions/__init__.py
@@ -6,11 +6,13 @@ into the :data:`~roscope.entities.expose.action_parse_methods` registry.
 
 from roscope.entities.actions import (  # noqa: F401
     arg,
+    composable_node_container,
     env,
     event_handler,
     executable,
     group,
     include,
+    load_composable_nodes,
     log,
     node,
     param,
