+++ roscope/entities/descriptions/__init__.py
@@ -1,0 +1,7 @@
+"""descriptions Module."""
+
+from .composable_node import ComposableNode
+
+__all__ = [
+    "ComposableNode",
+]
