+++ tests/test_renderer.py
@@ -2,15 +2,13 @@
 
 from __future__ import annotations
 
+from roscope.entities.actions.composable_node_container import ComposableNodeContainer
 from roscope.entities.actions.executable import ExecuteProcess
 from roscope.entities.actions.group import GroupAction
+from roscope.entities.actions.load_composable_nodes import LoadComposableNodes
 from roscope.entities.actions.marker import SourceMarker
-from roscope.entities.actions.node import (
-    ComposableNode,
-    ComposableNodeContainer,
-    LoadComposableNodes,
-    Node,
-)
+from roscope.entities.actions.node import Node
+from roscope.entities.descriptions import ComposableNode
 from roscope.entities.state import LaunchContext, ResolverState
 from roscope.renderer import render_resolved_xml
